Added emphasis on constraints in README file.
This update underscores the significance of constraints within method parameters. Not just in the example exercise but for all  methods requiring parameters. Constraints, such as length and content, play a crucial role in preventing bugs and handling various cases effectively. This clarification is essential, particularly in scenarios where the input's nature directly impacts the method's behaviour.