Update
Reading the previous exercise's text I got confused thinking that I should put inside the tree only words that appears only one time in the entire book. I think that text is cleaner this way, but it is only my opinion...

The changes about the code are made according to the exercise's text.