Update BankAccount  (Functional Programming)
The previous Comparator gives a wrong list if the difference of two or more accounts' amount is less that one (because of casting to integer).
In fact, if account1 has amount = 1000
and        account2 -> amount = 1000.4

the comparator does (int)1000 - (int)1000.4 = 1000 - 1000 = 0
So the two account looks the same, while they're not.
Comparator.comparingDouble prevents this.