Image update
As the last paragraph says "Dependency Inversion Principle is telling us is to invert the dependency. The flow of control will still follow the same path. However, now both our objects will depend on the abstraction level of the interface. Thus, ClassB inverts its dependency on ClassA" but the image intro-dependency-inversion-after.avif shows the opposite, Flow of dependencies is following the same path and is not inverted, while Flow of control, on the other hand, is inverted contrarily to what was previously stated. I have tried to upload an edited version of the image with the two captions swapped, but it won't let me add it to the slides/images folder in order to load it up for the change commitment.