Update RemoveFirstTwoCharsIf.java
I think there was a mistake in the previous version. If both the first two letters were to be deleted, the first and third letters were actually deleted, rather than the first and second. In the test it did not give an error because in Good the second and third letters are the same.