canBalance simplified sum calculation and 1 method
Instead of doing 2 methods and writing more code, I find this solution easier to understand and to solve this problem. 
I do the total sum of the vector, in a for cicle I calculate the left sum and then with subtraction i find the rigth sum, if them are equal means that the array can split.